PCI: hisi: Rearchitect the driver to enable multiple SoCs support
This patch re-architects the driver to support cleanly multiple SoCs:
- Now there is a structure of function pointers that is
  specific to each SoC
- Each SoC will implement its SoCs specific functions in
  a separate file

The patch has been tested on D02 with a Mellanox MCX312B-XCCT
It needs to be tested on Hip06 eval board
