+++ drivers/pci/host/pcie-hisi.c
@@ -1,10 +1,11 @@
 /*
- * PCIe host controller driver for HiSilicon Hip05 SoC
+ * PCIe host controllers common functions for HiSilicon SoCs drivers
  *
  * Copyright (C) 2015 HiSilicon Co., Ltd. http://www.hisilicon.com
  *
- * Author: Zhou Wang <wangzhou1@hisilicon.com>
- *         Dacai Zhu <zhudacai@hisilicon.com>
+ * Authors: Zhou Wang <wangzhou1@hisilicon.com>
+ *          Dacai Zhu <zhudacai@hisilicon.com>
+ *          Gabriele Paoloni <gabriele.paoloni@huawei.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 as
@@ -21,172 +22,13 @@
 #include <linux/of_device.h>
 
 #include "pcie-designware.h"
+#include "pcie-hisi.h"
 
-#define PCIE_HOST_HIP05                                 0x660
-#define PCIE_HOST_HIP06                                 0x1610
-
-#define PCIE_SUBCTRL_MODE_REG                           0x2800
-#define PCIE_SUBCTRL_SYS_STATE4_REG                     0x6818
-#define PCIE_SLV_DBI_MODE                               0x0
-#define PCIE_SLV_SYSCTRL_MODE                           0x1
-#define PCIE_SLV_CONTENT_MODE                           0x2
-#define PCIE_SLV_MSI_ASID                               0x10
-#define PCIE_LTSSM_LINKUP_STATE                         0x11
-#define PCIE_LTSSM_STATE_MASK                           0x3F
-#define PCIE_MSI_ASID_ENABLE                            (0x1 << 12)
-#define PCIE_MSI_ASID_VALUE                             (0x1 << 16)
-#define PCIE_MSI_TRANS_ENABLE                           (0x1 << 12)
-#define PCIE_MSI_TRANS_REG                              0x1c8
-#define PCIE_MSI_LOW_ADDRESS                            0x1b4
-#define PCIE_MSI_HIGH_ADDRESS                           0x1c4
-#define PCIE_GITS_TRANSLATER                            0x10040
-
-#define PCIE_SYS_CTRL20_REG                             0x20
-#define PCIE_RD_TAB_SEL                                 BIT(31)
-#define PCIE_RD_TAB_EN                                  BIT(30)
-#define PCIE_CFG_BAR0BASE                               0x10
-#define PCIE_DB2_ENABLE_SHIFT                           BIT(0)
-#define PCIE_DBI_CS2_ENABLE                             0x1
-#define PCIE_DBI_CS2_DISABLE                            0x0
-
-#define PCIE_CTRL_7_REG                                 0x114
-#define PCIE_LTSSM_ENABLE_SHIFT                         BIT(11)
-#define PCIE_SUBCTRL_RESET_REQ_REG                      0xA00
-#define PCIE0_2_SUBCTRL_RESET_REQ_REG(port_id) \
-	(PCIE_SUBCTRL_RESET_REQ_REG + (port_id << 3))
-#define PCIE3_SUBCTRL_RESET_REQ_REG                     0xA68
-
-#define PCIE_SUBCTRL_DRESET_REQ_REG                     0xA04
-#define PCIE0_2_SUBCTRL_DRESET_REQ_REG(port_id) \
-	(PCIE_SUBCTRL_DRESET_REQ_REG + (port_id << 3))
-#define PCIE3_SUBCTRL_DRESET_REQ_REG                    0xA6C
-
-#define PCIE_SUBCTRL_RESET_ST_REG                       0x5A00
-#define PCIE0_2_SUBCTRL_RESET_ST_REG(port_id) \
-	(PCIE_SUBCTRL_RESET_ST_REG + (port_id << 2))
-#define PCIE3_SUBCTRL_RESET_ST_REG                      0x5A34
-
-#define PCIE_SUBCTRL_SC_PCIE0_CLK_DIS_REG               0x304
-#define PCIE_SUBCTRL_SC_PCIE0_2_CLK_DIS_REG(port_id) \
-	(PCIE_SUBCTRL_SC_PCIE0_CLK_DIS_REG + (port_id << 3))
-#define PCIE_SUBCTRL_SC_PCIE3_CLK_DIS_REG               0x324
-
-#define PCIE_SUBCTRL_SC_PCIE0_CLK_ST_REG                0x5300
-#define PCIE_SUBCTRL_SC_PCIE0_2_CLK_ST_REG(port_id) \
-	(PCIE_SUBCTRL_SC_PCIE0_CLK_ST_REG + (port_id << 2))
-#define PCIE_SUBCTRL_SC_PCIE3_CLK_ST_REG                0x5310
-
-#define PCIE_SUBCTRL_SC_PCIE0_CLK_EN_REG                0x300
-#define PCIE_SUBCTRL_SC_PCIE0_2_CLK_EN_REG(port_id) \
-	(PCIE_SUBCTRL_SC_PCIE0_CLK_EN_REG + (port_id << 3))
-#define PCIE_SUBCTRL_SC_PCIE3_CLK_EN_REG                0x320
-
-#define PCIE_ASSERT_RESET_ON                            1
-#define PCIE_DEASSERT_RESET_ON                          0
-#define PCIE_CLOCK_ON                                   1
-#define PCIE_CLOCK_OFF                                  0
-
-#define PCIE_PCS_LOCAL_RESET_REQ                        0xAC0
-#define PCIE_PCS_LOCAL_DRESET_REQ                       0xAC4
-#define PCIE_PCS_RESET_REQ_REG                          0xA60
-#define PCIE_PCS_RESET_REG_ST                           0x5A30
-#define PCIE_PCS_LOCAL_DRESET_ST                        0x5A60
-#define PCIE_PCS_LOCAL_RESET_ST                         0x5A60
-#define PCIE_PCS_DRESET_REQ_REG                         0xA64
-#define PCIE_M_PCS_IN15_CFG                             0x74
-#define PCIE_M_PCS_IN13_CFG                             0x34
-#define PCIE_PCS_RXDETECTED                             0xE4
-#define PCIE_PCS_SERDES_STATUS                          0x8108
-#define HIP05_PCIE_CORE_RESET                           0x1
-#define HIP06_PCIE_CORE_RESET                           0x3
-#define PCIE_CORE_MODE_REG                              0xF8
-#define PCIE_LINK_WIDTH_SPEED_CONTROL                   0x80C
-#define PORT_LOGIC_SPEED_CHANGE                         (0x1 << 17)
-
-#define DS_API(lane) ((0x1FF6c + 8 * (15 - lane)) * 2)
-#define PCIE_DFE_ENABLE_VAL                             0x3851
-
-#define PCIE_RD_ERR_RESPONSE_EN                         BIT(31)
-#define PCIE_WR_ERR_RESPONSE_EN                         BIT(30)
-#define PCIE_PCS_APB_RESET_REQ                          0xAC8
-#define PCIE_PCS_APB_DRESET_REQ                         0xACC
-#define PCIE_CTRL_19_REG                                0x1c
-
-enum pcie_mac_phy_rate_e {
-	PCIE_GEN1 = 0,		/* PCIE 1.0 */
-	PCIE_GEN2 = 1,		/* PCIE 2.0 */
-	PCIE_GEN3 = 2		/* PCIE 3.0 */
-};
-
-#define to_hisi_pcie(x)	container_of(x, struct hisi_pcie, pp)
-
-struct hisi_pcie {
-	void __iomem *subctrl_base;
-	void __iomem *reg_base;
-	void __iomem *phy_base;
-	void __iomem *serdes_base;
-	void __iomem *ctrl_base;
-	u32 port_id;
-	struct pcie_port pp;
-	u64 msi_addr;
-	int soc_type;
-};
-
-static inline void hisi_pcie_subctrl_writel(struct hisi_pcie *pcie,
-					    u32 val, u32 reg)
-{
-	writel(val, pcie->subctrl_base + reg);
-}
-
-static inline u32 hisi_pcie_subctrl_readl(struct hisi_pcie *pcie, u32 reg)
-{
-	return readl(pcie->subctrl_base + reg);
-}
-
-static inline void hisi_pcie_apb_writel(struct hisi_pcie *pcie,
-					u32 val, u32 reg)
-{
-	writel(val, pcie->reg_base + reg);
-}
-
-static inline u32 hisi_pcie_apb_readl(struct hisi_pcie *pcie, u32 reg)
-{
-	return readl(pcie->reg_base + reg);
-}
-static inline void hisi_pcie_pcs_writel(struct hisi_pcie *pcie,
-					u32 val, u32 reg)
-{
-	writel(val, pcie->phy_base + reg);
-}
-
-static inline u32 hisi_pcie_pcs_readl(struct hisi_pcie *pcie, u32 reg)
-{
-	return readl(pcie->phy_base + reg);
-}
-
-static inline void hisi_pcie_serdes_writel(struct hisi_pcie *pcie,
-					   u32 val, u32 reg)
-{
-	writel(val, pcie->serdes_base + reg);
-}
-
-static inline void hisi_pcie_ctrl_writel(struct hisi_pcie *pcie, u32 val, u32 reg)
-{
-	return writel(val, pcie->ctrl_base + reg);
-}
-
-static inline u32 hisi_pcie_ctrl_readl(struct hisi_pcie *pcie, u32 reg)
-{
-	return readl(pcie->ctrl_base + reg);
-}
-
-#define PCIE_SYS_STATE4                                 0x31c
-#define PCIE_PCS_SERDES_STATUS_HIP06                    0x504
 /*
  * Change mode to indicate the same reg_base to base of PCIe host configure
  * registers, base of RC configure space or base of vmid/asid context table
  */
-static void hisi_pcie_change_apb_mode(struct hisi_pcie *pcie, u32 mode)
+void hisi_pcie_change_apb_mode(struct hisi_pcie *pcie, u32 mode)
 {
 	u32 val;
 	u32 bit_mask;
@@ -207,145 +49,6 @@ static void hisi_pcie_change_apb_mode(struct hisi_pcie *pcie, u32 mode)
 	hisi_pcie_subctrl_writel(pcie, val, reg);
 }
 
-/* Configure vmid/asid table in PCIe host */
-static void hisi_pcie_config_context(struct hisi_pcie *pcie)
-{
-	int i;
-	u32 val;
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		/*
-		 * enable to clean vmid and asid tables though apb bus
-		 * */
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_SYSCTRL_MODE);
-
-		val = hisi_pcie_apb_readl(pcie, PCIE_SYS_CTRL20_REG);
-		/* enable ar channel */
-		val |= PCIE_RD_TAB_SEL | PCIE_RD_TAB_EN;
-		hisi_pcie_apb_writel(pcie, val, PCIE_SYS_CTRL20_REG);
-
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_CONTENT_MODE);
-		for (i = 0; i < 0x800; i++)
-			hisi_pcie_apb_writel(pcie, 0x0, i * 4);
-
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_SYSCTRL_MODE);
-		/* enable aw channel */
-		val &= (~PCIE_RD_TAB_SEL);
-		val |= PCIE_RD_TAB_EN;
-		hisi_pcie_apb_writel(pcie, val, PCIE_SYS_CTRL20_REG);
-
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_CONTENT_MODE);
-
-		/*
-		 * init vmid and asid tables for all PCIe devices as 0
-		 * vmid table: 0 ~ 0x3ff, asid table: 0x400 ~ 0x7ff
-		 */
-		for (i = 0; i < 0x800; i++)
-			hisi_pcie_apb_writel(pcie, 0x0, i * 4);
-
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_SYSCTRL_MODE);
-
-		val = hisi_pcie_apb_readl(pcie, PCIE_SYS_CTRL20_REG);
-		/* disable ar channel */
-		val |= PCIE_RD_TAB_SEL;
-		val &= (~PCIE_RD_TAB_EN);
-		hisi_pcie_apb_writel(pcie, val, PCIE_SYS_CTRL20_REG);
-		/* disable aw channel */
-		val &= ((~PCIE_RD_TAB_SEL) & (~PCIE_RD_TAB_EN));
-		hisi_pcie_apb_writel(pcie, val, PCIE_SYS_CTRL20_REG);
-
-		hisi_pcie_apb_writel(pcie, pcie->msi_addr & 0xffffffff, PCIE_MSI_LOW_ADDRESS);
-		hisi_pcie_apb_writel(pcie, pcie->msi_addr >> 32, PCIE_MSI_HIGH_ADDRESS);
-		hisi_pcie_apb_writel(pcie, PCIE_MSI_ASID_ENABLE | PCIE_MSI_ASID_VALUE,
-				     PCIE_SLV_MSI_ASID);
-		hisi_pcie_apb_writel(pcie, PCIE_MSI_TRANS_ENABLE, PCIE_MSI_TRANS_REG);
-
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_DBI_MODE);
-	} else {
-		hisi_pcie_ctrl_writel(pcie, pcie->msi_addr & 0xffffffff, PCIE_MSI_LOW_ADDRESS);
-		hisi_pcie_ctrl_writel(pcie, pcie->msi_addr >> 32, PCIE_MSI_HIGH_ADDRESS);
-		val = hisi_pcie_ctrl_readl(pcie, PCIE_MSI_TRANS_REG);
-		val |= PCIE_MSI_TRANS_ENABLE;
-		hisi_pcie_ctrl_writel(pcie, val, PCIE_MSI_TRANS_REG);
-	}
-}
-
-static int hisi_pcie_link_up(struct pcie_port *pp)
-{
-	u32 val = 0;
-	struct hisi_pcie *hisi_pcie = to_hisi_pcie(pp);
-
-	if (hisi_pcie->soc_type == PCIE_HOST_HIP05)
-		val = hisi_pcie_subctrl_readl(hisi_pcie,
-				PCIE_SUBCTRL_SYS_STATE4_REG + 0x100 * hisi_pcie->port_id);
-	else if (hisi_pcie->soc_type == PCIE_HOST_HIP06)
-		val = hisi_pcie_ctrl_readl(hisi_pcie, PCIE_SYS_STATE4);
-
-	return ((val & PCIE_LTSSM_STATE_MASK) == PCIE_LTSSM_LINKUP_STATE);
-}
-
-static void hisi_pcie_enable_ltssm(struct hisi_pcie *pcie, bool on)
-{
-	u32 val;
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_SYSCTRL_MODE);
-		val = hisi_pcie_apb_readl(pcie, PCIE_CTRL_7_REG);
-		if (on)
-			val |= PCIE_LTSSM_ENABLE_SHIFT;
-		else
-			val &= ~(PCIE_LTSSM_ENABLE_SHIFT);
-		hisi_pcie_apb_writel(pcie, val, PCIE_CTRL_7_REG);
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_DBI_MODE);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		val = hisi_pcie_ctrl_readl(pcie, PCIE_CTRL_7_REG);
-		if (on)
-		/* In Hip06 host will fill all 1s in CFG cpu address space
-		 * when host sends a read CFG TLP with wrong BDF if we set
-		 * PCIE_RD_ERR_RESPONSE_EN as 1. And host will trigger a error
-		 * interrupt when host sends a read/write CFG TLP with wrong
-		 * BDF if we set PCIE_RD_ERR_RESPONSE_EN/PCIE_WR_ERR_RESPONSE_EN
-		 * as 1
-		 */
-			val |= PCIE_LTSSM_ENABLE_SHIFT |
-			       PCIE_RD_ERR_RESPONSE_EN |
-			       PCIE_WR_ERR_RESPONSE_EN;
-		else
-			val &= ~(PCIE_LTSSM_ENABLE_SHIFT);
-		hisi_pcie_ctrl_writel(pcie, val, PCIE_CTRL_7_REG);
-	}
-}
-
-static void hisi_pcie_gen3_dfe_enable(struct hisi_pcie *pcie)
-{
-	u32 val;
-	u32 lane;
-	u32 port_id = pcie->port_id;
-	u32 current_speed;
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		if (port_id == 3)
-			return;
-		val = hisi_pcie_subctrl_readl(pcie,
-					      PCIE_SUBCTRL_SYS_STATE4_REG +
-					      0x100 * port_id);
-		current_speed = (val >> 6) & 0x3;
-		if (current_speed != PCIE_GEN3)
-			return;
-		for (lane = 0; lane < 8; lane++)
-			hisi_pcie_serdes_writel(pcie,
-					PCIE_DFE_ENABLE_VAL, DS_API(lane) + 4);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		val = hisi_pcie_apb_readl(pcie, 0x31c);
-		current_speed = (val >> 6) & 0x3;
-		if (current_speed != PCIE_GEN3)
-			return;
-	}
-
-	dev_info(pcie->pp.dev, "enable DFE success!\n");
-}
-
-
 static void hisi_pcie_core_reset_ctrl(struct hisi_pcie *pcie, bool reset_on)
 {
 	u32 reg_reset_ctrl;
@@ -355,6 +58,7 @@ static void hisi_pcie_core_reset_ctrl(struct hisi_pcie *pcie, bool reset_on)
 	u32 reset_status_checked;
 	unsigned long timeout;
 	u32 port_id = pcie->port_id;
+	u32 pcie_pcs_core_reset = pcie->soc_ops->pcie_pcs_core_reset;
 
 	if (port_id == 3) {
 		reg_reset_ctrl = PCIE3_SUBCTRL_RESET_REQ_REG;
@@ -368,23 +72,14 @@ static void hisi_pcie_core_reset_ctrl(struct hisi_pcie *pcie, bool reset_on)
 
 	if (reset_on) {
 		/* if core is link up, stop the ltssm state machine first */
-		if (hisi_pcie_link_up(&pcie->pp))
-			hisi_pcie_enable_ltssm(pcie, 0);
+		if (pcie->soc_ops->hisi_pcie_link_up(pcie))
+			pcie->soc_ops->hisi_pcie_enable_ltssm(pcie, 0);
 
-		if (pcie->soc_type == PCIE_HOST_HIP05)
-			hisi_pcie_subctrl_writel(pcie, HIP05_PCIE_CORE_RESET,
-					reg_reset_ctrl);
-		else if (pcie->soc_type == PCIE_HOST_HIP06)
-			hisi_pcie_subctrl_writel(pcie, HIP06_PCIE_CORE_RESET,
-					reg_reset_ctrl);
-	} else {
-		if (pcie->soc_type == PCIE_HOST_HIP05)
-			hisi_pcie_subctrl_writel(pcie, HIP05_PCIE_CORE_RESET,
-					reg_dereset_ctrl);
-		else if (pcie->soc_type == PCIE_HOST_HIP06)
-			hisi_pcie_subctrl_writel(pcie, HIP06_PCIE_CORE_RESET,
-					reg_dereset_ctrl);
-	}
+		hisi_pcie_subctrl_writel(pcie, pcie_pcs_core_reset,
+				reg_reset_ctrl);
+	} else
+		hisi_pcie_subctrl_writel(pcie, pcie_pcs_core_reset,
+				reg_dereset_ctrl);
 
 	/* wait a delay for 1s */
 	timeout = jiffies + HZ * 1;
@@ -392,23 +87,11 @@ static void hisi_pcie_core_reset_ctrl(struct hisi_pcie *pcie, bool reset_on)
 	do {
 		reset_status = hisi_pcie_subctrl_readl(pcie, reg_reset_status);
 		if (reset_on) {
-			if (pcie->soc_type == PCIE_HOST_HIP05) {
-				reset_status &= HIP05_PCIE_CORE_RESET;
-				reset_status_checked =
-						(reset_status != HIP05_PCIE_CORE_RESET);
-			} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-				reset_status &= HIP06_PCIE_CORE_RESET;
-				reset_status_checked =
-						(reset_status != HIP06_PCIE_CORE_RESET);
-			}
+			reset_status &= pcie_pcs_core_reset;
+			reset_status_checked = (reset_status != pcie_pcs_core_reset);
 		} else {
-			if (pcie->soc_type == PCIE_HOST_HIP05) {
-				reset_status &= HIP05_PCIE_CORE_RESET;
-				reset_status_checked = (reset_status != 0);
-			} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-				reset_status &= HIP06_PCIE_CORE_RESET;
-				reset_status_checked = (reset_status != 0);
-			}
+			reset_status &= pcie_pcs_core_reset;
+			reset_status_checked = (reset_status != 0);
 		}
 	} while ((reset_status_checked) && (time_before(jiffies, timeout)));
 
@@ -426,6 +109,7 @@ static void hisi_pcie_clock_ctrl(struct hisi_pcie *pcie, bool clock_on)
 	u32 clock_status_checked;
 	unsigned long timeout;
 	u32 port_id = pcie->port_id;
+	u32 pcie_clk_ctrl = pcie->soc_ops->pcie_clk_ctrl;
 
 	if (port_id == 3) {
 		reg_clock_disable = PCIE_SUBCTRL_SC_PCIE3_CLK_DIS_REG;
@@ -438,31 +122,18 @@ static void hisi_pcie_clock_ctrl(struct hisi_pcie *pcie, bool clock_on)
 		reg_clock_status = PCIE_SUBCTRL_SC_PCIE0_2_CLK_ST_REG(port_id);
 	}
 
-	if (clock_on) {
-		if (pcie->soc_type == PCIE_HOST_HIP05)
-			hisi_pcie_subctrl_writel(pcie, 0x3, reg_clock_enable);
-		else if (pcie->soc_type == PCIE_HOST_HIP06)
-			hisi_pcie_subctrl_writel(pcie, 0x7, reg_clock_enable);
-	} else {
-		if (pcie->soc_type == PCIE_HOST_HIP05)
-			hisi_pcie_subctrl_writel(pcie, 0x3, reg_clock_disable);
-		else if (pcie->soc_type == PCIE_HOST_HIP06)
-			hisi_pcie_subctrl_writel(pcie, 0x7, reg_clock_disable);
-	}
+	if (clock_on)
+		hisi_pcie_subctrl_writel(pcie, pcie_clk_ctrl, reg_clock_enable);
+	else
+		hisi_pcie_subctrl_writel(pcie, pcie_clk_ctrl, reg_clock_disable);
 	timeout = jiffies + HZ * 1;
 	do {
 		clock_status = hisi_pcie_subctrl_readl(pcie, reg_clock_status);
-		if (clock_on) {
-			if (pcie->soc_type == PCIE_HOST_HIP05)
-				clock_status_checked = ((clock_status & 0x03) != 0x3);
-			else if (pcie->soc_type == PCIE_HOST_HIP06)
-				clock_status_checked = ((clock_status & 0x07) != 0x7);
-		} else {
-			if (pcie->soc_type == PCIE_HOST_HIP05)
-				clock_status_checked = ((clock_status & 0x03) != 0);
-			else if (pcie->soc_type == PCIE_HOST_HIP06)
-				clock_status_checked = ((clock_status & 0x07) != 0);
-		}
+		if (clock_on)
+			clock_status_checked =
+					((clock_status & pcie_clk_ctrl) != pcie_clk_ctrl);
+		else
+			clock_status_checked = ((clock_status & pcie_clk_ctrl) != 0);
 	} while ((clock_status_checked) && (time_before(jiffies, timeout)));
 
 	/* get a timeout error */
@@ -494,7 +165,7 @@ static void hisi_pcie_deassert_pcs_reset(struct hisi_pcie *pcie)
 	val = 1 << port_id;
 	hisi_pcie_subctrl_writel(pcie, val, PCIE_PCS_LOCAL_DRESET_REQ);
 
-	if (pcie->soc_type == PCIE_HOST_HIP06)
+	if (pcie->soc_ops->soc_type == PCIE_HOST_HIP06)
 		hisi_pcie_subctrl_writel(pcie, val, PCIE_PCS_APB_DRESET_REQ);
 
 	val = 0xff << (port_id << 3);
@@ -538,7 +209,7 @@ static void hisi_pcie_assert_pcs_reset(struct hisi_pcie *pcie)
 	reg = 1 << port_id;
 	hisi_pcie_subctrl_writel(pcie, reg, PCIE_PCS_LOCAL_RESET_REQ);
 
-	if (pcie->soc_type == PCIE_HOST_HIP06)
+	if (pcie->soc_ops->soc_type == PCIE_HOST_HIP06)
 		hisi_pcie_subctrl_writel(pcie, reg, PCIE_PCS_APB_RESET_REQ);
 
 	reg = 0xff << (port_id << 3);
@@ -571,56 +242,8 @@ static void hisi_pcie_assert_pcs_reset(struct hisi_pcie *pcie)
 		dev_err(pcie->pp.dev, "error:pcs assert reset failed\n");
 }
 
-static void hisi_pcie_init_pcs(struct hisi_pcie *pcie)
+void pcie_equalization_common(struct hisi_pcie *pcie)
 {
-	u32 i;
-	u32 lane_num = pcie->pp.lanes;
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		if (pcie->port_id <= 2) {
-			hisi_pcie_serdes_writel(pcie, 0x212, 0xc088);
-
-			hisi_pcie_pcs_writel(pcie, 0x2026044, 0x8020);
-			hisi_pcie_pcs_writel(pcie, 0x2126044, 0x8060);
-			hisi_pcie_pcs_writel(pcie, 0x2126044, 0x80c4);
-			hisi_pcie_pcs_writel(pcie, 0x2026044, 0x80e4);
-			hisi_pcie_pcs_writel(pcie, 0x4018, 0x80a0);
-			hisi_pcie_pcs_writel(pcie, 0x804018, 0x80a4);
-			hisi_pcie_pcs_writel(pcie, 0x11201100, 0x80c0);
-			hisi_pcie_pcs_writel(pcie, 0x3, 0x15c);
-			hisi_pcie_pcs_writel(pcie, 0x0, 0x158);
-		} else {
-			for (i = 0; i < lane_num; i++)
-				hisi_pcie_pcs_writel(pcie, 0x46e000,
-							 PCIE_M_PCS_IN15_CFG + (i << 2));
-			for (i = 0; i < lane_num; i++)
-				hisi_pcie_pcs_writel(pcie, 0x1001,
-							 PCIE_M_PCS_IN13_CFG + (i << 2));
-
-			hisi_pcie_pcs_writel(pcie, 0xffff, PCIE_PCS_RXDETECTED);
-		}
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		u32 val;
-
-		for (i = 0; i < lane_num; i++) {
-			val = hisi_pcie_pcs_readl(pcie, 0x204 + i * 0x4);
-			val |= (1u << 20);
-			hisi_pcie_pcs_writel(pcie, val, 0x204 + i * 0x4);
-		}
-		hisi_pcie_pcs_writel(pcie, 0x4e20, 0x264);
-	}
-}
-
-void pcie_equalization(struct hisi_pcie *pcie)
-{
-	u32 val;
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		hisi_pcie_apb_writel(pcie, 0x1400, 0x890);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		hisi_pcie_apb_writel(pcie, 0x1c00, 0x890);
-	}
-
 	hisi_pcie_apb_writel(pcie, 0xfd7, 0x894);
 
 	hisi_pcie_apb_writel(pcie, 0x0, 0x89c);
@@ -644,38 +267,6 @@ void pcie_equalization(struct hisi_pcie *pcie)
 	hisi_pcie_apb_writel(pcie, 0x9, 0x89c);
 	hisi_pcie_apb_writel(pcie, 0xd0b, 0x898);
 	hisi_pcie_apb_writel(pcie, 0x103ff21, 0x8a8);
-
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		val = hisi_pcie_apb_readl(pcie, 0x80);
-		val |= 0x80;
-		hisi_pcie_apb_writel(pcie, val, 0x80);
-
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x184);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x188);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x18c);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x190);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x164);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x168);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x16c);
-		hisi_pcie_apb_writel(pcie, 0x44444444, 0x170);
-
-		val = hisi_pcie_ctrl_readl(pcie, 0x2d0);
-		val &= (~0x3f);
-		val |= 0x5;
-		hisi_pcie_ctrl_writel(pcie, val, 0x2d0);
-	}
-}
-
-static void hisi_pcie_mode_set(struct hisi_pcie *pcie)
-{
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_SYSCTRL_MODE);
-		hisi_pcie_apb_writel(pcie, 0x4 << 28, PCIE_CORE_MODE_REG);
-		hisi_pcie_change_apb_mode(pcie, PCIE_SLV_DBI_MODE);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
-		hisi_pcie_ctrl_writel(pcie, 0x4 << 28, PCIE_CORE_MODE_REG);
-	}
 }
 
 static void hisi_pcie_spd_set(struct hisi_pcie *pcie, u32 spd)
@@ -702,18 +293,6 @@ static void hisi_pcie_spd_control(struct hisi_pcie *pcie)
 	hisi_pcie_apb_writel(pcie, val, PCIE_LINK_WIDTH_SPEED_CONTROL);
 }
 
-static void hisi_pcie_portnum_set(struct hisi_pcie *pcie, u32 num)
-{
-	u32 val;
-
-	if (pcie->soc_type == PCIE_HOST_HIP06) {
-		val = hisi_pcie_ctrl_readl(pcie, PCIE_CTRL_19_REG);
-		val &= ~(0xff);
-		val |= num;
-		hisi_pcie_ctrl_writel(pcie, val, PCIE_CTRL_19_REG);
-	}
-}
-
 void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 {
 	u32 val = 0;
@@ -727,12 +306,12 @@ void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 	hisi_pcie_assert_core_reset(pcie);
 	hisi_pcie_assert_pcs_reset(pcie);
 
-	if (pcie->soc_type == PCIE_HOST_HIP05) {
+	if (pcie->soc_ops->soc_type == PCIE_HOST_HIP05) {
 		/* de-assert phy reset */
 		hisi_pcie_deassert_pcs_reset(pcie);
 		/* de-assert core reset */
 		hisi_pcie_deassert_core_reset(pcie);
-	} else if (pcie->soc_type == PCIE_HOST_HIP06) {
+	} else if (pcie->soc_ops->soc_type == PCIE_HOST_HIP06) {
 		/* de-assert core reset */
 		hisi_pcie_deassert_core_reset(pcie);
 		/* de-assert phy reset */
@@ -743,10 +322,10 @@ void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 	hisi_pcie_clock_ctrl(pcie, PCIE_CLOCK_ON);
 
 	/* initialize phy */
-	hisi_pcie_init_pcs(pcie);
+	pcie->soc_ops->hisi_pcie_init_pcs(pcie);
 
 	/* set controller to RC mode */
-	hisi_pcie_mode_set(pcie);
+	pcie->soc_ops->hisi_pcie_mode_set(pcie);
 
 	/* set target link speed */
 	hisi_pcie_spd_set(pcie, 3);
@@ -757,7 +336,8 @@ void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 	 * establish ITS table for PCIe devices. However, PCIe host send
 	 * port_id + request_id + MSI_vector to ITS TRANSLATION register
 	 */
-	hisi_pcie_portnum_set(pcie, 0);
+	if (pcie->soc_ops->hisi_pcie_portnum_set)
+		pcie->soc_ops->hisi_pcie_portnum_set(pcie, 0);
 
 	/* set link speed control*/
 	hisi_pcie_spd_control(pcie);
@@ -765,21 +345,18 @@ void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 	/* setup root complex */
 	dw_pcie_setup_rc(&pcie->pp);
 
-	pcie_equalization(pcie);
+	pcie->soc_ops->pcie_equalization(pcie);
 
 	/* assert LTSSM enable */
-	hisi_pcie_enable_ltssm(pcie, 1);
+	pcie->soc_ops->hisi_pcie_enable_ltssm(pcie, 1);
 
 	/* check if the link is up or not */
 	while (!dw_pcie_link_up(&pcie->pp)) {
+		u32 pcs_serdes_status = pcie->soc_ops->pcs_serdes_status;
 		mdelay(100);
 		count++;
 		if (count == 10) {
-			if (pcie->soc_type == PCIE_HOST_HIP05)
-				val = hisi_pcie_pcs_readl(pcie, PCIE_PCS_SERDES_STATUS);
-			else if (pcie->soc_type == PCIE_HOST_HIP06)
-				val = hisi_pcie_pcs_readl(pcie,
-						PCIE_PCS_SERDES_STATUS_HIP06);
+			val = hisi_pcie_pcs_readl(pcie, pcs_serdes_status);
 
 			dev_info(pcie->pp.dev,
 				 "PCIe Link Failed! PLL Locked: 0x%x\n", val);
@@ -790,7 +367,8 @@ void hisi_pcie_establish_link(struct hisi_pcie *pcie)
 	dev_info(pcie->pp.dev, "Link up\n");
 
 	/* dfe enable is just for 660 */
-	hisi_pcie_gen3_dfe_enable(pcie);
+	if (pcie->soc_ops->hisi_pcie_gen3_dfe_enable)
+		pcie->soc_ops->hisi_pcie_gen3_dfe_enable(pcie);
 	/*
 	 * add a 1s delay between linkup and enumeration, make sure
 	 * the EP device's configuration registers are prepared well
@@ -857,7 +435,7 @@ static void hisi_pcie_host_init(struct pcie_port *pp)
 	struct hisi_pcie *pcie = to_hisi_pcie(pp);
 
 	hisi_pcie_establish_link(pcie);
-	hisi_pcie_config_context(pcie);
+	pcie->soc_ops->hisi_pcie_config_context(pcie);
 	/*
 	 * The default size of BAR0 in p660 host bridge is 0x10000000,
 	 * which will bring problem when most resource has been allocated
@@ -891,6 +469,12 @@ static int hisi_pcie_cfg_write(struct pcie_port *pp, int where, int  size,
 		return PCIBIOS_BAD_REGISTER_NUMBER;
 
 	return PCIBIOS_SUCCESSFUL;
+}
+
+int hisi_pcie_link_up(struct pcie_port *pp)
+{
+	struct hisi_pcie *hisi_pcie = to_hisi_pcie(pp);
+	return hisi_pcie->soc_ops->hisi_pcie_link_up(hisi_pcie);
 }
 
 static struct pcie_host_ops hisi_pcie_host_ops = {
@@ -928,19 +512,7 @@ static int hisi_add_pcie_port(struct pcie_port *pp,
 	return 0;
 }
 
-static const struct of_device_id hisi_pcie_of_match[] = {
-	{
-		.compatible = "hisilicon,hip05-pcie",
-		.data = (void *) PCIE_HOST_HIP05,
-	},
-	{
-		.compatible = "hisilicon,hip06-pcie",
-		.data = (void *) PCIE_HOST_HIP06,
-	},
-	{},
-};
-
-static int hisi_pcie_probe(struct platform_device *pdev)
+int hisi_pcie_probe(struct platform_device *pdev)
 {
 	struct hisi_pcie *hisi_pcie;
 	struct pcie_port *pp;
@@ -948,6 +520,7 @@ static int hisi_pcie_probe(struct platform_device *pdev)
 	struct resource *reg;
 	struct resource *subctrl;
 	struct resource *phy;
+	struct device_driver *driver;
 	int ret;
 
 	hisi_pcie = devm_kzalloc(&pdev->dev, sizeof(*hisi_pcie), GFP_KERNEL);
@@ -956,9 +529,10 @@ static int hisi_pcie_probe(struct platform_device *pdev)
 
 	pp = &hisi_pcie->pp;
 	pp->dev = &pdev->dev;
+	driver = (pdev->dev).driver;
 
-	match = of_match_device(hisi_pcie_of_match, &pdev->dev);
-	hisi_pcie->soc_type = (int) match->data;
+	match = of_match_device(driver->of_match_table, &pdev->dev);
+	hisi_pcie->soc_ops = (struct pcie_soc_ops *) match->data;
 
 	subctrl = platform_get_resource_byname(pdev, IORESOURCE_MEM, "subctrl");
 	hisi_pcie->subctrl_base = devm_ioremap_nocache(&pdev->dev,
@@ -984,7 +558,7 @@ static int hisi_pcie_probe(struct platform_device *pdev)
 		return PTR_ERR(hisi_pcie->phy_base);
 	}
 
-	if (hisi_pcie->soc_type == PCIE_HOST_HIP05) {
+	if (hisi_pcie->soc_ops->soc_type == PCIE_HOST_HIP05) {
 		struct resource *serdes;
 		serdes = platform_get_resource_byname(pdev, IORESOURCE_MEM, "serdes");
 		hisi_pcie->serdes_base = devm_ioremap_resource(&pdev->dev, serdes);
@@ -992,12 +566,11 @@ static int hisi_pcie_probe(struct platform_device *pdev)
 			dev_err(pp->dev, "cannot get serdes base\n");
 			return PTR_ERR(hisi_pcie->serdes_base);
 		}
-	}
-
-	if (hisi_pcie->soc_type == PCIE_HOST_HIP06)
+		hisi_pcie->ctrl_base = NULL;
+	} else if (hisi_pcie->soc_ops->soc_type == PCIE_HOST_HIP06)
 		hisi_pcie->ctrl_base = hisi_pcie->reg_base + 0x1000;
 	else
-		hisi_pcie->ctrl_base = NULL;
+		dev_err(pp->dev, "unsopported soc type\n");
 
 	ret = hisi_add_pcie_port(pp, pdev);
 	if (ret)
@@ -1007,22 +580,3 @@ static int hisi_pcie_probe(struct platform_device *pdev)
 
 	return 0;
 }
-
-MODULE_DEVICE_TABLE(of, hisi_pcie_of_match);
-
-static struct platform_driver hisi_pcie_driver = {
-	.driver = {
-		   .name = "hisi-pcie",
-		   .of_match_table = hisi_pcie_of_match,
-	},
-};
-
-static int __init hisi_pcie_init(void)
-{
-	return platform_driver_probe(&hisi_pcie_driver, hisi_pcie_probe);
-}
-subsys_initcall(hisi_pcie_init);
-
-MODULE_AUTHOR("Zhou Wang <wangzhou1@hisilicon.com>");
-MODULE_AUTHOR("Dacai Zhu <zhudacai@hisilicon.com>");
-MODULE_LICENSE("GPL v2");
